backend/ats_api_updated/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
PyPDF2
python-docx
google-generativeai
pydantic-settings
python-dotenv
